refactor(cli-startup-prompt): extract startup prompt functions
Move startup prompt related functions from `exec_strategy.rs` to a new `startup_prompt.rs` module.

- `print_terminal_safe_stderr`
- `prompt_startup_termination`
- `StartupPromptTerminalGuard`
- `prompt_startup_termination_for_child`

This refactoring centralizes the logic for handling startup-blocked processes, improving modularity and code organization.
